+++ src/main/scala/ot/dispatcher/kafka/context/JobsContainer.scala
@@ -3,13 +3,10 @@ package ot.dispatcher.kafka.context
 import play.api.libs.json.JsValue
 
 import java.util.concurrent.ConcurrentLinkedQueue
-import scala.collection.mutable.ArrayBuffer
 
-object CommandsContainer {
-
-  var values: ArrayBuffer[JsValue] = ArrayBuffer[JsValue]()
-
+/**
+ * Contains thread-safe queue for receiving jobs from Kafka broker and processing them in app working cycle
+ */
+object JobsContainer {
   var syncValues: ConcurrentLinkedQueue[JsValue] = new ConcurrentLinkedQueue[JsValue]()
-
-  var changedValues: ArrayBuffer[JsValue] = new ArrayBuffer[JsValue]()
 }